requirements.txt
psycopg2-binary
docker-compose
celery==4.3.0
redis==3.3.5
django_celery_beat==1.5.0
djangorestframework==3.10.2
django_celery_results==1.1.2
djangorestframework_simplejwt